Add files via upload
The project is employed  to predict epileptic
seizure in EEG signals using some selected Ensemble Machine
Learning models. XGBoost, Random Forest, Decision Tree, Ada
Boost, Gradient Boost, and Bagging were the Ensemble Machine
Learning models used in this study. This research focused on
three classification problems from the Bonn EEG Dataset from
Bonn University in Germany. There were five responses of EEG
signals from the dataset, which are: seizure activity (S), tumour
affected area (F), tumour healthy area (N), eyes closed (O), and
eyes opened (Z). The first classification was Z-S, the second
was O-F, and lastly, the third was N-S. Data was acquired
from the Clinique of Epileptology of Bonn University. Principal
Component Analysis (PCA) was used for feature extraction, and
Standard Scaler was used to scale the data’s features. Features
of the extracted data were classified using XG Boost, Random
Forest, Gradient Boost, Ada Boost, Bagging and Decision Tree
models. The first classification was Z-S, the second was OF, and lastly, the third was N-S. The model prediction of
epileptic seizure was done, and then evaluation of the model was
carried out using the 10-fold cross-validation. The performance
of the classifier was assessed in terms of accuracy, precision,
recall, and F1 measure. The Random Forest model performed
better than the rest of the models in terms of accuracy in
the first classification problem, Z-S, and the third classification
problem, N-S. XGBoost model performed better in terms of
accuracy in the second classification problem, O-F. Based on
the result above, the Random Forest model was chosen for
the model deployment. Random forest was used to deploy all
three classification problems, and it was deployed in Flask Web
Application Framework. The success of the Random Forest model
in this research work suggests its clinical applicability in the
early detection of epileptic seizures. This knowledge is crucial
for healthcare professionals, neurologists, and researchers seeking
robust tools to improve patient care and outcomes in the field of
neurology